Massive maps update
Finished plotting all waredas on google maps. Managed to get properties parsed out of the kml and into the topo json. Reduced object sizes from 50+mb in geojson to 1mb in topojson. Managed to get info windows to appear and the color of each wareda to be dependent on a specific property.